pyproject.toml
[project]
name = "nonebot-plugin-chatgpt-api"
version = "1.0.0"
description = "NoneBot2 plugin for AI chat with OpenAI API."
authors = [
    { name = "SanJerry007", email = "sandy375408563@gmail.com" },
]
license = { text = "MIT" }
dependencies = ["nonebot2>=2.0.0rc2", "nonebot-adapter-onebot>=2.1.5", "nonebot-plugin-apscheduler>=0.2.0", "nonebot-plugin-htmlrender>=0.2.0.1", "openai", "pandas"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/SanJerry007/nonebot-plugin-chatgpt-api"
Repository = "https://github.com/SanJerry007/nonebot-plugin-chatgpt-api"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
